Plugin WIP

+++ DVD/SRC/GCM.h
@@ -1,5 +1,5 @@
 // externals for DVD callbacks (see DVD.h)
 BOOL    GCMSelectFile(char *file);
-void    GCMSeek(long position);
-void    GCMRead(u8 *buf, long length);
+void    GCMSeek(int position);
+void    GCMRead(uint8_t *buf, int length);
 void    GCMClose();
